requirements.txt
Flask==3.1.2
huggingface-hub==0.35.3
pillow==11.3.0
psycopg2-binary==2.9.10
